Move code to avoid circular dep
